[cleanup] Moved control cost to core folder
