+++ benchmark/arm-manipulation-timings.cpp
@@ -19,7 +19,7 @@
 #include "crocoddyl/core/costs/cost-sum.hpp"
 #include "crocoddyl/multibody/costs/frame-placement.hpp"
 #include "crocoddyl/multibody/costs/state.hpp"
-#include "crocoddyl/multibody/costs/control.hpp"
+#include "crocoddyl/core/costs/control.hpp"
 #include "crocoddyl/multibody/actuations/full.hpp"
 
 #include "crocoddyl/core/utils/callbacks.hpp"
