+++ benchmark/factory/arm-kinova.hpp
@@ -22,7 +22,7 @@
 #include "crocoddyl/core/costs/cost-sum.hpp"
 #include "crocoddyl/multibody/costs/frame-placement.hpp"
 #include "crocoddyl/multibody/costs/state.hpp"
-#include "crocoddyl/multibody/costs/control.hpp"
+#include "crocoddyl/core/costs/control.hpp"
 
 namespace crocoddyl {
 namespace benchmark {
